move ShardingConfigurationEventBusEvent
